backend/requirements.txt
google-genai
fastapi
uvicorn
python-multipart
python-dotenv
httpx
